Fix updater

+++ themes/btcsumo/update-all-feeds.php
@@ -11,6 +11,8 @@ define( 'BTCSUMO_FEED_UPDATE_LIMIT', MINUTE_IN_SECONDS );
 // How far back do we go with feed items?
 define( 'BTCSUMO_FEED_KEEP_LIMIT', 30 * DAY_IN_SECONDS );
 
+// Reset all the feed items?
+$reset = isset( $_GET['reset'] );
 
 // Get all the active feeds.
 $feed_posts = get_posts( [
@@ -26,7 +28,7 @@ foreach ( $feed_posts as $feed_post ) {
   $feed_error  = get_post_meta( $feed_post->ID, 'feed-error',       true );
 
   // Limit the fetch. Faulty feeds skip 1 turn.
-  if ( $last_update > ( time() - BTCSUMO_FEED_UPDATE_LIMIT ) || $feed_error ) {
+  if ( ! $reset && $last_update > ( time() - BTCSUMO_FEED_UPDATE_LIMIT ) || $feed_error ) {
     delete_post_meta( $feed_post->ID, 'feed-error', $feed_error );
     continue;
   }
@@ -43,12 +45,12 @@ foreach ( $feed_posts as $feed_post ) {
   if ( ! is_wp_error( $feed_loaded ) ) {
 
     // Get the current list of feed items.
-    $feed_items = get_post_meta( $feed_post->ID, 'feed-feed-items', true );
+    $feed_items = ( $reset ) ? [] : get_post_meta( $feed_post->ID, 'feed-feed-items', true );
+    $first_update = false;
     if ( empty( $feed_items ) ) {
       $feed_items = [];
+      $first_update = true;
     }
-
-    $new_feed_item_count = 0;
 
     // Loop through all the fetched feed items and save the new ones to the current list.
     foreach ( $feed_loaded->get_items() as $feed_item ) {
@@ -59,9 +61,8 @@ foreach ( $feed_posts as $feed_post ) {
       ];
 
       // Is this one new?
-      if ( empty( $feed_items ) || count( $feed_items ) === $new_feed_item_count || $feed_items[0]->permalink !== $new_feed_item->permalink ) {
+      if ( $first_update || $feed_items[0]->permalink !== $new_feed_item->permalink ) {
         $feed_items[] = $new_feed_item;
-        $new_feed_item_count++;
       } else {
         // No new entries.
         break;
@@ -92,8 +93,8 @@ foreach ( $feed_posts as $feed_post ) {
     delete_post_meta( $feed_post->ID, 'feed-error', $feed_error );
 
   } else {
-    // Feed is faulty.
-    update_post_meta( $feed_post->ID, 'feed-error', true );
+    // Feed is faulty, remember the error message.
+    update_post_meta( $feed_post->ID, 'feed-error', $feed_loaded->get_error_message() );
   }
 
   // Update the time of the current feed update.
